pyproject.toml
[tool.poetry]
name = "тестовое задание в"
version = "0.1.0"
description = ""
authors = ["posholokia <ilya.posholokk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
punq = "^0.7.0"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
gunicorn = "^22.0.0"
uvicorn = "^0.30.3"
asyncpg = "^0.29.0"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
isort = "^5.13.2"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
exclude = [
    ".ruff_cache",
    ".venv",
    "venv",
    "src/core/database/base.py",
    ]

line-length = 79
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
ignore = ["E712"]

[tool.isort]
order_by_type = true
force_sort_within_sections=false
multi_line_output=3
include_trailing_comma=true
line_length=79
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
length_sort_straight=true
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["core", ]
skip_glob=[
    ".git",
    ".idea",
    ".ci",
    ".ruff_cache",
    "venv",
    "src/core/database/base.py"
]
default_section = "THIRDPARTY"
known_fastapi=["fastapi", "starlette", "sqlalchemy"]
known_firstparty=["src"]

sections=[
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FASTAPI",
    "FIRSTPARTY",
    "LOCALFOLDER",
]